Single threading required for ML2 cisco/nexus driver
During multi-VM creation (creates neutron ports) debug indicates that
multiple threads are used in calls to the update and delete (on removing
the VMs) methods. Since database access is used in the cisco nexus MD
event methods for determining the current switch configuration, failures
are seen (on both updates and deletes) on programming the nexus switch.

This commit forces single threaded access to these methods.

Change-Id: I275d1a8e43d69c2218b87f80283d325456d8948a