ML2 Cisco Nexus Driver to handle switch restart
If the Nexus switch restarts, it no longer contains configuration
set by OpenStack ML2 Nexus Driver.  This changeset spawns a thread
which periodically sends 'get version' requests to the Nexus switch
for the purpose of monitoring the health of the underlying
ncclient/ssh connection.  The thread takes note of a lost connection.
When the connection transition back up, all configuration will be
replayed to the Nexus switch.

Upon OpenStack reset, all connections are assumed down.  When
connections are re-established, their configuration will be replayed.

This feature is configurable.  At this time, it is disabled(0) by
default.  To turn it on, set switch_heartbeat_time = 30 beneath
[ml2_cisco] of the local.conf file in devstack.

Change-Id: I9a8b6ed34a6b45093e23d5bf6fab641b0c27b1c6
Closes-bug:  #1407987