Error/Traceback during VM deletion; fail to delete VM
Issue: Error/Traceback during VM deletion, fails to delete VM,
the VM is stuck in "error" state and cannot be deleted.

Root cause is the way ML2 handles the exception to the
None segment. Changing it to not raise an exception,
as we want the delete to proceed. Will log an warning
message.

Change-Id: I5279fb01268de80ff85b6b4af5a6d18970a95099
Closes-bug: #1414209